Add fix for CVE-2025-26465
A vulnerability was found in OpenSSH when the VerifyHostKeyDNS option
is enabled. A machine-in-the-middle attack can be performed by a
malicious machine impersonating a legit server.
This issue occurs due to how OpenSSH mishandles err codes in specific
conditions when verifying the host key.
For an attack to be considered successful, attacker needs to manage
to exhaust the client's memory resource first,
turning the attack complexity high.

In the patch:
- Added error codes

Signed-off-by: Lucas RAVAGNIER <ravagnierlucas@gmail.com>